corrected menu_background
will ask for permission to change the names of these death screen backgrounds so that they are ranked and used accordingly to display how much the player progressed before they died. the more they progress, the more of an injured face they see in death screen. the smiling face, the neutral face with nose bleeding, neutral face with eye bleeding and broken berret and lastly completely broken and blood spillied berret with both eyes and mout bleeding.