READ DESCRIPTION OF THIS PATCH PLEASE
added zombie attack animation frames, they are complete now and all added to the project. the order of the animation is 3-2-4-1-4-2-3. before the start of these frames and at the end of the animation i want the zombi to get into its static pose and if its not alerted by anything such as players getting into its trigger boxes or if it doesnt die i want it to always stay in the static pose which is also added to the project with the previous patch. you can make adjustments such as zombi staying a bit longer in attack1 frame to signal the players its going to complete its attack so that the players can avoid the completion of the attack by slowing down. maybe zombi can also stay at attack2 image for longer to make it seem like it did a punchy, strong hit etc. i am leaving this to the developers taste, i gave little advices and the order of frames.