added buttons for menus, READ DESCRIPTION
the text inside the black version of the button must be full white. the black button is normal state of the button, the red version is when the button is selected or when the mouse cursor is on the button and green version of the button is when we click the button. when we click the button the buttons image should change to the green version and half a second later we get to the next menu or whatever else the button does. THE TEXT INSIDE THE BUTTON MUST CHANGE COLOR TO FULL BLACK WHEN THE BUTTON CHANGES INTO RED OR GREEN VERSIONS. DURING THE BUTTONS BLACK STATE, THE TEXT INSIDE IT MUST BE WHITE.